+++ Hisab/src/main/java/com/ANP/repository/CustomerDAO.java
@@ -117,7 +117,7 @@ public class CustomerDAO {
             cus.setOrgId(rs.getLong("customer.orgid"));
             cus.setState(rs.getString("customer.state"));
             cus.setSendPaymentReminders(rs.getBoolean("customer.sendPaymentReminders"));
-            cus.setAccountBalance(rs.getFloat("account.currentbalance"));
+            cus.setAccountBalance(rs.getBigDecimal("account.currentbalance"));
             cus.setCreateDate(rs.getTimestamp("customer.createdate"));
             cus.setCreatedbyId(rs.getString("customer.createdbyid"));
             cus.setAccountId(rs.getLong("account.id"));
